app/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.4.0
Pillow==10.3.0
gunicorn==22.0.0
djangorestframework-simplejwt
djoser
django-filter
# whitenoise